Deployed - install flow
