+++ components/MyLayout.js
@@ -1,7 +1,7 @@
 import Header from './Header'
 
 const layoutStyle = {
-  margin: 20,
+  margin: 0,
   padding: 20,
   border: '1px solid #DDD'
 }
