+++ now.json
@@ -1,0 +1,6 @@
+{
+  "version": 2,
+  "builds": [
+    { "src": "package.json", "use": "@now/next" }
+  ]
+}
