+++ package.json
@@ -7,7 +7,8 @@
     "test": "echo \"Error: no test specified\" && exit 1",
     "dev": "node server.js",
     "build": "next build",
-    "start": "next start"
+    "start": "next start",
+    "flow": "flow"
   },
   "keywords": [],
   "author": "",
@@ -27,5 +28,8 @@
     "react": "^16.7.0",
     "react-dom": "^16.7.0",
     "react-markdown": "^4.0.6"
+  },
+  "devDependencies": {
+    "flow-bin": "^0.93.0"
   }
 }
