+++ pages/index.js
@@ -1,5 +1,6 @@
 import Layout from '../components/MyLayout.js'
 import Link from 'next/link'
+import PostLink from '../components/PostLink.js'
 
 function getPosts () {
   return [
@@ -9,41 +10,18 @@ function getPosts () {
   ]
 }
 
-const PostLink = ({ post }) => (
-  <li>
-    <Link as={`/p/${post.id}`} href={`/post?title=${post.title}`}>
-      <a>{post.title}</a>
-    </Link>
-    <style jsx>{`
-      li {
-        list-style: none;
-        margin: 5px 0;
-      }
-
-      a {
-        text-decoration: none;
-        color: blue;
-        font-family: "Arial";
-      }
-
-      a:hover {
-        opacity: 0.6;
-      }
-    `}</style>
-  </li>
-)
-
 export default () => (
   <Layout>
-    <h1>My Blog</h1>
+    <h1>WHITESPACE</h1>
     <ul>
       {getPosts().map((post) => (
         <PostLink key={post.id} post={post}/>
       ))}
     </ul>
     <style jsx>{`
-      h1, a {
-        font-family: "Arial";
+      @import url('https://fonts.googleapis.com/css?family=Lato|Montserrat');
+      h1 {
+        font-family: "Montserrat";
       }
 
       ul {
