+++ server.js
@@ -1,30 +1,32 @@
-const express = require('express');
-const next = require('next');
+const express = require("express");
+const next = require("next");
 
-const dev = process.env.NODE_ENV != 'production';
+const dev = process.env.NODE_ENV != "production";
 const app = next({ dev });
-const handle = app.getRequestHandler()
+const handle = app.getRequestHandler();
 
-app.prepare()
-.then(() => {
-  const server = express();
+app
+  .prepare()
+  .then(() => {
+    const server = express();
+    console.log("server: " + server);
 
-  server.get('/p/:id', (req, res) => {
-    const actualPage = '/post';
-    const queryParams = { id: req.params.id };
-    app.render(req, res, actualPage, queryParams);
-  });
+    server.get("/p/:id", (req, res) => {
+      const actualPage = "/post";
+      const queryParams = { id: req.params.id };
+      app.render(req, res, actualPage, queryParams);
+    });
 
-  server.get('*', (req, res) => {
-    return handle(req, res);
-  });
+    server.get("*", (req, res) => {
+      return handle(req, res);
+    });
 
-  server.listen(3000, (err) => {
-    if (err) throw err;
-    console.log('> Ready on http://localhost:300');
+    server.listen(3000, err => {
+      if (err) throw err;
+      console.log("> Ready on http://localhost:3000");
+    });
   })
-})
-.catch((ex) => {
-  console.error(ex.stack);
-  process.exit(1);
-})
+  .catch(ex => {
+    console.error(ex.stack);
+    process.exit(1);
+  });
